requirements.txt
streamlit>=1.32.0
python-docx>=0.8.11
PyPDF2>=3.0.1
requests>=2.31.0

# Optional: AI-powered compliance analysis
langchain>=0.2.0
langchain-community>=0.2.0
openai>=1.12.0
faiss-cpu>=1.7.4
